Refactors training scripts
