+++ espnet/asr/asr_utils.py
@@ -3,35 +3,50 @@
 # Copyright 2017 Johns Hopkins University (Shinji Watanabe)
 #  Apache 2.0  (http://www.apache.org/licenses/LICENSE-2.0)
 
-import copy
 import json
 import logging
-# matplotlib related
-import os
-import shutil
-import tempfile
 
 # chainer related
 import chainer
 
 from chainer import training
-from chainer.training import extension
+from chainer.training import extensions
 
-from chainer.serializers.npz import DictionarySerializer
-from chainer.serializers.npz import NpzDeserializer
-
+import torch
 
 # io related
 import kaldi_io_py
-import matplotlib
-import numpy as np
-import torch
 
+import numpy as np
+
+from espnet.utils.pytorch_utils import torch_load
+from espnet.utils.pytorch_utils import torch_resume
+from espnet.utils.pytorch_utils import torch_save
+
+from espnet.utils.training.train_utils import add_attention_report
+from espnet.utils.training.train_utils import add_early_stop
+from espnet.utils.training.train_utils import add_tensorboard
+from espnet.utils.training.train_utils import REPORT_INTERVAL
+
+# matplotlib related
+import matplotlib
 
 matplotlib.use('Agg')
 
 
 # * -------------------- training iterator related -------------------- *
+
+def make_args_batchset(data, args):
+    """Make batch set from json dictionary
+
+    :param data: dictionary loaded from data.json
+    :param args: the program arguments
+    :return: list of batches
+    """
+    return make_batchset(data, args.batch_size, args.maxlen_in, args.maxlen_out, args.num_batches,
+                         args.ngpu if args.ngpu > 1 else 1)
+
+
 def make_batchset(data, batch_size, max_length_in, max_length_out,
                   num_batches=0, min_batch_size=1):
     """Make batch set from json dictionary
@@ -159,82 +174,6 @@ class CompareValueTrigger(object):
         self._summary = chainer.reporter.DictSummary()
 
 
-class PlotAttentionReport(extension.Extension):
-    """Plot attention reporter
-
-    :param function att_vis_fn: function of attention visualization
-    :param list data: list json utt key items
-    :param str outdir: directory to save figures
-    :param CustomConverter converter: function to convert data
-    :param int | torch.device device: device
-    :param bool reverse: If True, input and output length are reversed
-    """
-
-    def __init__(self, att_vis_fn, data, outdir, converter, device, reverse=False):
-        self.att_vis_fn = att_vis_fn
-        self.data = copy.deepcopy(data)
-        self.outdir = outdir
-        self.converter = converter
-        self.device = device
-        self.reverse = reverse
-        if not os.path.exists(self.outdir):
-            os.makedirs(self.outdir)
-
-    def __call__(self, trainer):
-        att_ws = self.get_attention_weights()
-        for idx, att_w in enumerate(att_ws):
-            filename = "%s/%s.ep.{.updater.epoch}.png" % (
-                self.outdir, self.data[idx][0])
-            att_w = self.get_attention_weight(idx, att_w)
-            self._plot_and_save_attention(att_w, filename.format(trainer))
-
-    def log_attentions(self, logger, step):
-        att_ws = self.get_attention_weights()
-        for idx, att_w in enumerate(att_ws):
-            att_w = self.get_attention_weight(idx, att_w)
-            plot = self.draw_attention_plot(att_w)
-            logger.add_figure("%s" % (self.data[idx][0]), plot.gcf(), step)
-            plot.clf()
-
-    def get_attention_weights(self):
-        batch = self.converter([self.converter.transform(self.data)], self.device)
-        att_ws = self.att_vis_fn(*batch)
-        return att_ws
-
-    def get_attention_weight(self, idx, att_w):
-        if self.reverse:
-            dec_len = int(self.data[idx][1]['input'][0]['shape'][0])
-            enc_len = int(self.data[idx][1]['output'][0]['shape'][0])
-        else:
-            dec_len = int(self.data[idx][1]['output'][0]['shape'][0])
-            enc_len = int(self.data[idx][1]['input'][0]['shape'][0])
-        if len(att_w.shape) == 3:
-            att_w = att_w[:, :dec_len, :enc_len]
-        else:
-            att_w = att_w[:dec_len, :enc_len]
-        return att_w
-
-    def draw_attention_plot(self, att_w):
-        import matplotlib.pyplot as plt
-        if len(att_w.shape) == 3:
-            for h, aw in enumerate(att_w, 1):
-                plt.subplot(1, len(att_w), h)
-                plt.imshow(aw, aspect="auto")
-                plt.xlabel("Encoder Index")
-                plt.ylabel("Decoder Index")
-        else:
-            plt.imshow(att_w, aspect="auto")
-            plt.xlabel("Encoder Index")
-            plt.ylabel("Decoder Index")
-        plt.tight_layout()
-        return plt
-
-    def _plot_and_save_attention(self, att_w, filename):
-        plt = self.draw_attention_plot(att_w)
-        plt.savefig(filename)
-        plt.close()
-
-
 def restore_snapshot(model, snapshot, load_fn=chainer.serializers.load_npz):
     """Extension to restore snapshot"""
 
@@ -272,179 +211,6 @@ def _adadelta_eps_decay(trainer, eps_decay):
         for p in optimizer.param_groups:
             p["eps"] *= eps_decay
             logging.info('adadelta eps decayed to ' + str(p["eps"]))
-
-
-def torch_snapshot(savefun=torch.save,
-                   filename='snapshot.ep.{.updater.epoch}'):
-    """Returns a trainer extension to take snapshots of the trainer for pytorch."""
-
-    @extension.make_extension(trigger=(1, 'epoch'), priority=-100)
-    def torch_snapshot(trainer):
-        _torch_snapshot_object(trainer, trainer, filename.format(trainer), savefun)
-
-    return torch_snapshot
-
-
-def _torch_snapshot_object(trainer, target, filename, savefun):
-    # make snapshot_dict dictionary
-    s = DictionarySerializer()
-    s.save(trainer)
-    if hasattr(trainer.updater.model, "model"):
-        # (for TTS)
-        if hasattr(trainer.updater.model.model, "module"):
-            model_state_dict = trainer.updater.model.model.module.state_dict()
-        else:
-            model_state_dict = trainer.updater.model.model.state_dict()
-    else:
-        # (for ASR)
-        if hasattr(trainer.updater.model, "module"):
-            model_state_dict = trainer.updater.model.module.state_dict()
-        else:
-            model_state_dict = trainer.updater.model.state_dict()
-    snapshot_dict = {
-        "trainer": s.target,
-        "model": model_state_dict,
-        "optimizer": trainer.updater.get_optimizer('main').state_dict()
-    }
-
-    # save snapshot dictionary
-    fn = filename.format(trainer)
-    prefix = 'tmp' + fn
-    tmpdir = tempfile.mkdtemp(prefix=prefix, dir=trainer.out)
-    tmppath = os.path.join(tmpdir, fn)
-    try:
-        savefun(snapshot_dict, tmppath)
-        shutil.move(tmppath, os.path.join(trainer.out, fn))
-    finally:
-        shutil.rmtree(tmpdir)
-
-
-# * -------------------- general -------------------- *
-class AttributeDict(object):
-    def __init__(self, obj):
-        self.obj = obj
-
-    def __getstate__(self):
-        return self.obj.items()
-
-    def __setstate__(self, items):
-        if not hasattr(self, 'obj'):
-            self.obj = {}
-        for key, val in items:
-            self.obj[key] = val
-
-    def __getattr__(self, name):
-        if name in self.obj:
-            return self.obj.get(name)
-        else:
-            return None
-
-    def __getitem__(self, name):
-        return self.obj[name]
-
-    def __len__(self):
-        return len(self.obj)
-
-    def fields(self):
-        return self.obj
-
-    def items(self):
-        return self.obj.items()
-
-    def keys(self):
-        return self.obj.keys()
-
-
-def get_model_conf(model_path, conf_path=None):
-    """Get model config information by reading a model config file (model.json)
-
-    :param str model_path: model path
-    :param str conf_path: optional model config path
-    """
-
-    if conf_path is None:
-        model_conf = os.path.dirname(model_path) + '/model.json'
-    else:
-        model_conf = conf_path
-    with open(model_conf, "rb") as f:
-        logging.info('reading a config file from ' + model_conf)
-        return json.load(f, object_hook=AttributeDict)
-
-
-def chainer_load(path, model):
-    """Function to load chainer model parameters
-
-    :param str path: model file or snapshot file to be loaded
-    :param chainer.Chain model: chainer model
-    """
-    if 'snapshot' in path:
-        chainer.serializers.load_npz(path, model, path='updater/model:main/')
-    else:
-        chainer.serializers.load_npz(path, model)
-
-
-def torch_save(path, model):
-    """Function to save torch model states
-
-    :param str path: file path to be saved
-    :param torch.nn.Module model: torch model
-    """
-    if hasattr(model, 'module'):
-        torch.save(model.module.state_dict(), path)
-    else:
-        torch.save(model.state_dict(), path)
-
-
-def torch_load(path, model):
-    """Function to load torch model states
-
-    :param str path: model file or snapshot file to be loaded
-    :param torch.nn.Module model: torch model
-    """
-    if 'snapshot' in path:
-        model_state_dict = torch.load(path, map_location=lambda storage, loc: storage)['model']
-    else:
-        model_state_dict = torch.load(path, map_location=lambda storage, loc: storage)
-    if hasattr(model, 'module'):
-        model.module.load_state_dict(model_state_dict)
-    else:
-        model.load_state_dict(model_state_dict)
-
-    del model_state_dict
-
-
-def torch_resume(snapshot_path, trainer):
-    """Function to resume from snapshot for pytorch
-
-    :param str snapshot_path: snapshot file path
-    :param instance trainer: chainer trainer instance
-    """
-    # load snapshot
-    snapshot_dict = torch.load(snapshot_path, map_location=lambda storage, loc: storage)
-
-    # restore trainer states
-    d = NpzDeserializer(snapshot_dict['trainer'])
-    d.load(trainer)
-
-    # restore model states
-    if hasattr(trainer.updater.model, "model"):
-        # (for TTS model)
-        if hasattr(trainer.updater.model.model, "module"):
-            trainer.updater.model.model.module.load_state_dict(snapshot_dict['model'])
-        else:
-            trainer.updater.model.model.load_state_dict(snapshot_dict['model'])
-    else:
-        # (for ASR model)
-        if hasattr(trainer.updater.model, "module"):
-            trainer.updater.model.module.load_state_dict(snapshot_dict['model'])
-        else:
-            trainer.updater.model.load_state_dict(snapshot_dict['model'])
-
-    # retore optimizer states
-    trainer.updater.get_optimizer('main').load_state_dict(snapshot_dict['optimizer'])
-
-    # delete opened snapshot
-    del snapshot_dict
 
 
 # * ------------------ recognition related ------------------ *
@@ -508,3 +274,196 @@ def add_results_to_json(js, nbest_hyps, char_list):
             logging.info('prediction : %s' % out_dic['rec_text'])
 
     return new_js
+
+
+def add_epsilon_decay(trainer, model, args):
+    """Adds the extension managing the epsilon decay
+
+    :param trainer: The trainer to add the extension to
+    :param model: The model to train
+    :param args: The program arguments
+    """
+    mtl_mode = get_mtl_mode(args.mtlalpha)
+    load_fn = chainer.serializers.load_npz if args.backend == 'chainer' else torch_load
+    if args.opt == 'adadelta':
+        if args.criterion == 'acc' and mtl_mode is not 'ctc':
+            trainer.extend(restore_snapshot(model, args.outdir + '/model.acc.best', load_fn=load_fn),
+                           trigger=CompareValueTrigger(
+                               'validation/main/acc',
+                               lambda best_value, current_value: best_value > current_value))
+            trainer.extend(adadelta_eps_decay(args.eps_decay),
+                           trigger=CompareValueTrigger(
+                               'validation/main/acc',
+                               lambda best_value, current_value: best_value > current_value))
+        elif args.criterion == 'loss':
+            trainer.extend(restore_snapshot(model, args.outdir + '/model.loss.best', load_fn=load_fn),
+                           trigger=CompareValueTrigger(
+                               'validation/main/loss',
+                               lambda best_value, current_value: best_value < current_value))
+            trainer.extend(adadelta_eps_decay(args.eps_decay),
+                           trigger=CompareValueTrigger(
+                               'validation/main/loss',
+                               lambda best_value, current_value: best_value < current_value))
+
+
+def add_snapshot(trainer, model, mtl_mode, savefun):
+    """Adds the model snapshot extension
+
+    :param trainer: The trainer to add the extension to
+    :param model: The model to save
+    :param mtl_mode: The mtl mode
+    :param savefun: The save function to use
+    """
+    trainer.extend(extensions.snapshot_object(model, 'model.loss.best', savefun=savefun),
+                   trigger=training.triggers.MinValueTrigger('validation/main/loss'))
+    if mtl_mode is not 'ctc':
+        trainer.extend(extensions.snapshot_object(model, 'model.acc.best', savefun=savefun),
+                       trigger=training.triggers.MaxValueTrigger('validation/main/acc'))
+
+
+def add_plot_report(trainer):
+    """Adds the plot report extension
+
+    :param trainer: The trainer to add the extension to
+    """
+    # Make a plot for training and validation values
+    trainer.extend(extensions.PlotReport(['main/loss', 'validation/main/loss',
+                                          'main/loss_ctc', 'validation/main/loss_ctc',
+                                          'main/loss_att', 'validation/main/loss_att'],
+                                         'epoch', file_name='loss.png'))
+    trainer.extend(extensions.PlotReport(['main/acc', 'validation/main/acc'],
+                                         'epoch', file_name='acc.png'))
+
+
+def get_dimensions(json_file):
+    """Returns the input and output dimensions given a json filepath
+
+    :param json_file: The json file containing the utterances data
+    :return: input dim, output dim
+    """
+    with open(json_file, 'rb') as f:
+        parsed_json = json.load(f)['utts']
+    utts = list(parsed_json.keys())
+    idim = int(parsed_json[utts[0]]['input'][0]['shape'][1])
+    odim = int(parsed_json[utts[0]]['output'][0]['shape'][1])
+    logging.info('#input dims : ' + str(idim))
+    logging.info('#output dims: ' + str(odim))
+    return idim, odim
+
+
+def get_mtl_mode(mtlalpha):
+    """Returns the mtl mode given the mtlalpha value
+
+    :param float mtlalpha: the multitask learning mode coefficient
+    :return: a string representing the mtlmode
+    """
+    # specify attention, CTC, hybrid mode
+    if mtlalpha == 1.0:
+        mtl_mode = 'ctc'
+        logging.info('Pure CTC mode')
+    elif mtlalpha == 0.0:
+        mtl_mode = 'att'
+        logging.info('Pure attention mode')
+    else:
+        mtl_mode = 'mtl'
+        logging.info('Multitask learning mode')
+    return mtl_mode
+
+
+def _get_trainer_lambda(trainer, is_pytorch):
+    return trainer.updater.get_optimizer('main').param_groups[0][
+        'eps'] if is_pytorch else trainer.updater.get_optimizer('main').eps
+
+
+def add_progress_report(trainer, args):
+    """Adds the logging and progress report extensions
+
+    :param trainer: The trainer to add the extensions to
+    :param args: The program arguments
+    """
+    is_pytorch = args.backend == 'pytorch'
+    trainer.extend(extensions.LogReport(trigger=(REPORT_INTERVAL, 'iteration')))
+    report_keys = ['epoch', 'iteration', 'main/loss', 'main/loss_ctc', 'main/loss_att',
+                   'validation/main/loss', 'validation/main/loss_ctc', 'validation/main/loss_att',
+                   'main/acc', 'validation/main/acc', 'elapsed_time']
+    if args.opt == 'adadelta':
+        trainer.extend(extensions.observe_value('eps', lambda trainer: _get_trainer_lambda(trainer, is_pytorch)),
+                       trigger=(REPORT_INTERVAL, 'iteration'))
+        report_keys.append('eps')
+    if is_pytorch:
+        if args.report_cer:
+            report_keys.append('validation/main/cer')
+        if args.report_wer:
+            report_keys.append('validation/main/wer')
+    trainer.extend(extensions.PrintReport(
+        report_keys), trigger=(REPORT_INTERVAL, 'iteration'))
+
+    trainer.extend(extensions.ProgressBar(update_interval=REPORT_INTERVAL))
+
+
+def prepare_trainer(updater, evaluator, converter, model, valid_json, args, device):
+    """Instantiates and adds common extensions to the trainer
+
+    :param updater: The training updater
+    :param evaluator: The training evaluator
+    :param converter: The batch converter
+    :param model: The model to train
+    :param valid_json: The validation json
+    :param args: The program arguments
+    :param device: The device to use
+    :return: the trainer
+    """
+    is_chainer = args.backend == 'chainer'
+    mtl_mode = get_mtl_mode(args.mtlalpha)
+    savefun = chainer.serializers.save_npz if is_chainer else torch_save
+    resume_fun = chainer.serializers.load_npz if is_chainer else torch_resume
+
+    trainer = training.Trainer(
+        updater, (args.epochs, 'epoch'), out=args.outdir)
+
+    if args.resume:
+        logging.info('resumed from %s' % args.resume)
+        resume_fun(args.resume, trainer)
+
+    trainer.extend(evaluator)
+    add_progress_report(trainer, args)
+    add_plot_report(trainer)
+    add_epsilon_decay(trainer, model, args)
+    add_snapshot(trainer, model, mtl_mode, savefun)
+    att_reporter = add_attention_report(trainer, model, args, valid_json, converter, device)
+    add_early_stop(trainer, args)
+    add_tensorboard(trainer, args.tensorboard_dir, att_reporter)
+    return trainer
+
+
+def single_beam_search(model, js, args, train_args, rnnlm):
+    """Executes single (no batch) beam search decoding
+
+    :param model: The model used for recognition
+    :param js: The input features as a json
+    :param args: The decoding arguments
+    :param train_args: The model training arguments
+    :param rnnlm: The RNNLM used for Shallow Fusion
+    :return: The json with the predictions
+    """
+    is_pytorch = args.backend == 'pytorch'
+    func = torch.no_grad if is_pytorch else chainer.no_backprop_mode
+    new_js = {}
+    with func():
+        for idx, name in enumerate(js.keys(), 1):
+            logging.info('(%d/%d) decoding ' + name, idx, len(js.keys()))
+            feat = kaldi_io_py.read_mat(js[name]['input'][0]['feat'])
+            nbest_hyps = model.recognize(feat, args, train_args.char_list, rnnlm)
+            new_js[name] = add_results_to_json(js[name], nbest_hyps, train_args.char_list)
+    return new_js
+
+
+def write_results(js, result_label):
+    """Writes the json dictionary to a file
+
+    :param js: the json object
+    :param str result_label: the file to write to
+    """
+    # TODO(watanabe) fix character coding problems when saving it
+    with open(result_label, 'wb') as f:
+        f.write(json.dumps({'utts': js}, indent=4, sort_keys=True).encode('utf_8'))
