+++ espnet/utils/pytorch_utils.py
@@ -1,0 +1,118 @@
+import os
+import shutil
+import tempfile
+
+import torch
+
+from chainer.serializers.npz import DictionarySerializer
+from chainer.serializers.npz import NpzDeserializer
+from chainer.training import extension
+
+
+def torch_save(path, model):
+    """Function to save torch model states
+
+    :param str path: file path to be saved
+    :param torch.nn.Module model: torch model
+    """
+    if hasattr(model, 'module'):
+        torch.save(model.module.state_dict(), path)
+    else:
+        torch.save(model.state_dict(), path)
+
+
+def torch_load(path, model):
+    """Function to load torch model states
+
+    :param str path: model file or snapshot file to be loaded
+    :param torch.nn.Module model: torch model
+    """
+    if 'snapshot' in path:
+        model_state_dict = torch.load(path, map_location=lambda storage, loc: storage)['model']
+    else:
+        model_state_dict = torch.load(path, map_location=lambda storage, loc: storage)
+    if hasattr(model, 'module'):
+        model.module.load_state_dict(model_state_dict)
+    else:
+        model.load_state_dict(model_state_dict)
+
+    del model_state_dict
+
+
+def torch_resume(snapshot_path, trainer):
+    """Function to resume from snapshot for pytorch
+
+    :param str snapshot_path: snapshot file path
+    :param instance trainer: chainer trainer instance
+    """
+    # load snapshot
+    snapshot_dict = torch.load(snapshot_path, map_location=lambda storage, loc: storage)
+
+    # restore trainer states
+    d = NpzDeserializer(snapshot_dict['trainer'])
+    d.load(trainer)
+
+    # restore model states
+    if hasattr(trainer.updater.model, "model"):
+        # (for TTS model)
+        if hasattr(trainer.updater.model.model, "module"):
+            trainer.updater.model.model.module.load_state_dict(snapshot_dict['model'])
+        else:
+            trainer.updater.model.model.load_state_dict(snapshot_dict['model'])
+    else:
+        # (for ASR model)
+        if hasattr(trainer.updater.model, "module"):
+            trainer.updater.model.module.load_state_dict(snapshot_dict['model'])
+        else:
+            trainer.updater.model.load_state_dict(snapshot_dict['model'])
+
+    # retore optimizer states
+    trainer.updater.get_optimizer('main').load_state_dict(snapshot_dict['optimizer'])
+
+    # delete opened snapshot
+    del snapshot_dict
+
+
+def torch_snapshot(savefun=torch.save,
+                   filename='snapshot.ep.{.updater.epoch}'):
+    """Returns a trainer extension to take snapshots of the trainer for pytorch."""
+
+    @extension.make_extension(trigger=(1, 'epoch'), priority=-100)
+    def torch_snapshot(trainer):
+        _torch_snapshot_object(trainer, filename.format(trainer), savefun)
+
+    return torch_snapshot
+
+
+def _torch_snapshot_object(trainer, filename, savefun):
+    # make snapshot_dict dictionary
+    s = DictionarySerializer()
+    s.save(trainer)
+    if hasattr(trainer.updater.model, "model"):
+        # (for TTS)
+        if hasattr(trainer.updater.model.model, "module"):
+            model_state_dict = trainer.updater.model.model.module.state_dict()
+        else:
+            model_state_dict = trainer.updater.model.model.state_dict()
+    else:
+        # (for ASR)
+        if hasattr(trainer.updater.model, "module"):
+            model_state_dict = trainer.updater.model.module.state_dict()
+        else:
+            model_state_dict = trainer.updater.model.state_dict()
+    snapshot_dict = {
+        "trainer": s.target,
+        "model": model_state_dict,
+        "optimizer": trainer.updater.get_optimizer('main').state_dict()
+    }
+
+    # save snapshot dictionary
+    fn = filename.format(trainer)
+    prefix = 'tmp' + fn
+    tmpdir = tempfile.mkdtemp(prefix=prefix, dir=trainer.out)
+    tmppath = os.path.join(tmpdir, fn)
+    try:
+        savefun(snapshot_dict, tmppath)
+        shutil.move(tmppath, os.path.join(trainer.out, fn))
+    finally:
+        shutil.rmtree(tmpdir)
